Changed look of ping scan Added lower limit for ping interval Some small wording changes in Display.cs Fixed conversion error in PingResults

+++ src/PowerPing/PingResults.cs
@@ -67,7 +67,7 @@ namespace PowerPing
                 MinTime = time;
 
             // Work out average
-            sum += time;
+            sum += (long) time;
             AvgTime = (double) sum / Received; // Avg = Total / Count
 
             CurTime = time;
